Reworked barrows (this now supports a barrows casket to choose 3 barrows items from, uses new area system aswell as tb player for 1 minute after completion) Fixed an issue where npcs would walk diagonally even if frozen Added support for selling items to the shops

+++ src/main/java/com/rs/java/game/player/actions/combat/Magic.java
@@ -867,6 +867,7 @@ public class Magic {
 			player.getPackets().sendGameMessage("You cannot teleport out of here.");
 			return false;
 		}
+		if (player.isLocked()) return false;
 		player.getInventory().deleteItem(itemId, 1);
 		int lockTime = 5;
 		player.lock(lockTime);
